+++ examples/combo.py
@@ -4,7 +4,7 @@ def selected(value):
     print(value)
 
 app = App()
-combo = Combo(app, ["Nothing", "Something"], command = selected)
-combo.add_option("Everything")
+combo = Combo(app, ["Nothing", "Something"], command = selected, selected="Something")
+combo.append("Everything")
 combo2 = Combo(app, ["hi", "bye"])
 app.display()