+++ tests/test_slider.py
@@ -8,7 +8,7 @@ from common_test import (
     display_test,
     text_test,
     color_test, 
-    size_text_test,
+    size_pixel_test,
     events_test,
     cascaded_properties_test,
     inherited_properties_test
@@ -169,7 +169,7 @@ def test_color():
 def test_size():
     a = App()
     s = Slider(a)
-    size_text_test(s)
+    size_pixel_test(s)
     a.destroy()
 
 def test_events():